random_face_detection/requirements.txt
lancedb
PyYAML
scikit-image
degirum
degirum-tools